requirements.txt
# =============================================================================
# The Life Shield - Production Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# --- Database ---
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9

# --- Validation & Serialization ---
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# --- HTTP Client ---
httpx==0.27.0
aiohttp==3.9.5

# --- Task Queue ---
celery==5.4.0
redis==5.0.4
flower==2.0.1

# --- Caching ---
hiredis==2.3.2

# --- AWS ---
boto3==1.34.96
botocore==1.34.96

# --- Payment ---
stripe==9.6.0

# --- Communication ---
twilio==9.0.5
sendgrid==6.11.0

# --- AI APIs ---
openai==1.30.1
anthropic==0.26.0

# --- Utilities ---
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0
phonenumbers==8.13.37
python-dateutil==2.9.0
pytz==2024.1
cryptography==42.0.7

# --- Monitoring ---
sentry-sdk[fastapi]==2.3.1

# --- CORS & Security ---
python-jose==3.3.0
